+++ lib/std/set.lua
@@ -149,7 +149,7 @@ end
 -- local Set = require "std.set".prototype
 -- assert (std.type (Set) == "Set")
 prototype = Container {
-  _type = "Set",
+  _type = "std.set.Set",
 
   --- Set object initialisation.
   --
@@ -240,8 +240,34 @@ prototype = Container {
     for k in pairs (self) do
       keys[#keys + 1] = tostring (k)
     end
-    table.sort (keys)
-    return type (self) .. " {" .. table.concat (keys, ", ") .. "}"
+    _sort (keys)
+    return type (self) .. " {" .. _concat (keys, ", ") .. "}"
+  end,
+
+  --- Return a loadable serialization of this object, where possible.
+  -- @function prototype:__pickle
+  -- @treturn string pickled object representation
+  -- @see std.string.pickle
+  __pickle = function (self)
+    local mt, keys = getmetatable (self), {}
+    for k in pairs (self) do
+      keys[#keys + 1] = pickle (k)
+    end
+    _sort (keys)
+    if _type (mt._module) == "string" then
+      -- object with _module set
+      return _concat {
+	'require "',
+	mt._module,
+	'".prototype {',
+	_concat (keys, ","),
+	"}",
+      }
+    end
+    -- rely on caller preloading `local ObjName = require "module".prototype`
+    return _concat {
+      mt._type, " {", _concat (keys, ","), "}"
+    }
   end,
 }